requirements.txt
django==5.2
mysqlclient==2.2.7
social-auth-app-django==5.4.3
python-dotenv==1.0.1
djangorestframework==3.15.0
reportlab==4.1.0 